// lib/api/generated/model/conversation.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { ConversationLastMessage } from './conversationLastMessage';
import type { ConversationOther } from './conversationOther';

export interface Conversation {
  matchId: string;
  createdAt: string;
  other: ConversationOther;
  lastMessage: ConversationLastMessage | null;
  unreadCount: number;
}